+++ pages/admin/index.js
@@ -1,26 +1,15 @@
-import { DeleteOutlined, EditOutlined } from "@ant-design/icons";
 import { Area } from "@ant-design/plots";
-import {
-	Col,
-	Divider,
-	message,
-	Popconfirm,
-	Row,
-	Select,
-	Space,
-	Table,
-	Tooltip,
-} from "antd";
+import { Col, Row, Select, Space } from "antd";
 import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { API_BASE_URL } from "../../apiconstants";
+import OrderTable, { orderData } from "../../components/admin/orderTable";
 import AdminLayout from "../../components/layouts/adminLayout";
 import useProvider from "../../hooks/useProvider";
 
 const { Option } = Select;
 
 const Dashboard = () => {
-	const [data, setData] = useState(null);
 	const {
 		state: {
 			user: { accessToken },
@@ -73,172 +62,16 @@ const Dashboard = () => {
 			.catch((e) => console.log(e));
 	}, [accessToken]);
 
-	const deleteOrder = (order) => {
-		axios
-			.delete(`${API_BASE_URL}/order/${order._id}`, {
-				headers: {
-					token: `Bearer ${accessToken}`,
-				},
-			})
-			.then((res) => {
-				message.success(res.data.message);
-				setData(data.filter((item) => item != order));
-			})
-			.catch((e) => console.log(e));
-	};
-
-	const handleStatus = (value, field) => {
-		axios
-			.put(
-				`${API_BASE_URL}/order/${field.id}`,
-				{ status: value },
-				{
-					headers: {
-						token: `Bearer ${accessToken}`,
-					},
-				}
-			)
-			.then((res) => {
-				message.success(res.data.message);
-			})
-			.catch(() => message.success("Failed to change status"));
-	};
-
-	const columns = [
-		{
-			title: "Order id",
-			dataIndex: "_id",
-			key: "2",
-			width: 100,
-			render: (id) => <Tooltip title={id}>#{id.slice(15)}</Tooltip>,
-		},
-		{
-			title: "Customer Name",
-			dataIndex: "user_name",
-			key: "name",
-			width: 200,
-		},
-		{
-			title: "Customer Email",
-			dataIndex: "email",
-			key: "email",
-			width: 250,
-		},
-		{
-			title: "Phone",
-			dataIndex: "phone",
-			key: "3",
-			width: 200,
-		},
-		{
-			title: "Address",
-			dataIndex: "address",
-			key: "4",
-			width: 280,
-		},
-		{
-			title: "Total Price",
-			dataIndex: "total_price",
-			key: "5",
-			width: 120,
-			defaultSortOrder: "descend",
-			sorter: (a, b) => a.total_price - b.total_price,
-		},
-		{
-			title: "Status",
-			key: "6",
-			width: 150,
-			render: (order) => (
-				<Select
-					style={{ width: 150 }}
-					defaultValue={order.status.toLowerCase()}
-					onChange={(value, field) => handleStatus(value, field)}
-				>
-					<Option id={order._id} value="pending">
-						<div className="flex items-center">
-							<div className="m-1 mr-2 w-2 h-2 relative rounded-full bg-yellow-400" />
-							Pending
-						</div>
-					</Option>
-					<Option id={order._id} value="approved">
-						<div className="flex items-center">
-							<div className="m-1 mr-2 w-2 h-2 relative rounded-full bg-lime-300" />
-							Approved
-						</div>
-					</Option>
-					<Option id={order._id} value="shifted">
-						<div className="flex items-center">
-							<div className="m-1 mr-2 w-2 h-2 relative rounded-full bg-violet-500" />
-							Shifted
-						</div>
-					</Option>
-					<Option id={order._id} value="completed">
-						<div className="flex items-center">
-							<div className="m-1 mr-2 w-2 h-2 relative rounded-full bg-green-500" />
-							Completed
-						</div>
-					</Option>
-					<Option id={order._id} value="cancled">
-						<div className="flex items-center">
-							<div className="m-1 mr-2 w-2 h-2 relative rounded-full bg-stone-300" />
-							Cancled
-						</div>
-					</Option>
-					<Option id={order._id} value="rejected">
-						<div className="flex items-center">
-							<div className="m-1 mr-2 w-2 h-2 relative rounded-full bg-red-500" />
-							Rejected
-						</div>
-					</Option>
-				</Select>
-			),
-			filters: [
-				{
-					text: "Pending",
-					value: "pending",
-				},
-				{
-					text: "Approved",
-					value: "approved",
-				},
-				{
-					text: "Shifted",
-					value: "shifted",
-				},
-				{
-					text: "Completed",
-					value: "completed",
-				},
-				{
-					text: "Cancled",
-					value: "cancled",
-				},
-				{
-					text: "Rejected",
-					value: "rejected",
-				},
-			],
-			onFilter: (value, record) => record.status.indexOf(value) === 0,
-		},
-		{
-			title: "Actions",
-			key: "actions",
-			width: 80,
-			render: (order) => (
-				<Space split={<Divider type="vertical" />}>
-					<Popconfirm
-						title="Are you sure you want to delete this order?"
-						onConfirm={() => deleteOrder(order)}
-						okText="Yes"
-						cancelText="No"
-						placement="topRight"
-					>
-						<DeleteOutlined />
-					</Popconfirm>
-					<EditOutlined />
-				</Space>
-			),
-		},
+	const date = [
+		"2022-04-25T18:12:25.742Z",
+		"2022-04-25T16:56:32.425Z",
+		"2022-04-21T17:55:26.476Z",
+		"2022-04-20T20:23:05.779Z",
+		"2022-04-20T20:20:15.675Z",
+		"2022-04-19T20:05:14.465Z",
+		"2022-04-09T21:31:30.113Z",
+		"2022-04-15T05:30:38.200Z",
+		"2022-04-19T21:02:00.717Z",
 	];
 
 	return (
@@ -311,7 +144,9 @@ const Dashboard = () => {
 							}}
 						>
 							<Col>Total Orders</Col>
-							<Col className="text-xl">{data?.length || 0}</Col>
+							<Col className="text-xl">
+								{orderData?.length || 0}
+							</Col>
 						</Row>
 					</Col>
 					<Col xs={24} md={12} lg={6}>
@@ -328,13 +163,8 @@ const Dashboard = () => {
 						</Row>
 					</Col>
 				</Row>
-				<Table
-					columns={columns}
-					dataSource={data}
-					scroll={{ x: 1550 }}
-					pagination={{ position: ["bottomCenter"] }}
-					size="small"
-				/>
+
+				<OrderTable />
 			</Space>
 		</AdminLayout>
 	);
